Move frontend tests
